Add monthly bills functionality
- Monthly bills are permanent, and are designed for bills that repeat
indefinetly every month.

- At the start of every month all monthly bills appear as unpaid,
the user can then set the bill price and pay. At the moment a monthly
bill is paid, a corresponding bill is created, with the same name,
billing company and value set by the user. The bill payment date is set
as the date the user paid the monthly bill